:tractor: Added TypeScript & Restructured
