+++ Cargo.toml
@@ -1,21 +1,23 @@
 [package]
 name = "game-of-life"
 version = "0.1.0"
-authors = ["jul1u5 <marozas.julius@gmail.com>"]
+authors = ["Julius Marozas <marozas.julius@gmail.com>"]
+
+description = "Conway's Game of Life implemented in WebAssembly"
+repository = "https://github.com/jul1u5/game-of-life"
+license = "MIT"
 
 [lib]
-crate-type = ["cdylib"]
+crate-type = ["cdylib", "rlib"]
+path = "crate/src/lib.rs"
 
 [features]
-default = ["console_error_panic_hook", "wee_alloc"]
-
-[profile.release]
-lto = true
+default-features = ["console_error_panic_hook"]
 
 [dependencies]
 cfg-if = "0.1.6"
 wasm-bindgen = "0.2"
-js-sys = "0.3.6"
+js-sys = "0.3.14"
 bit-vec = "0.5.0"
 
 # The `console_error_panic_hook` crate provides better debugging of panics by
@@ -27,4 +29,19 @@ console_error_panic_hook = { version = "0.1.1", optional = true }
 # `wee_alloc` is a tiny allocator for wasm that is only ~1K in code size
 # compared to the default allocator's ~10K. It is slower than the default
 # allocator, however.
-wee_alloc = { version = "0.4.1", optional = true }
+#
+# Unfortunately, `wee_alloc` requires nightly Rust when targeting wasm for now.
+wee_alloc = { version = "0.4.2", optional = true }
+
+[dev-dependencies]
+wasm-bindgen-test = "0.2"
+
+[dependencies.web-sys]
+version = "0.3"
+features = [
+  "console",
+]
+
+# [profile.release]
+# opt-level = 'z'
+# lto = true
