+++ LICENSE
@@ -1,4 +1,4 @@
-Copyright (c) 2018 jul1u5 <marozas.julius@gmail.com>
+Copyright (c) 2019 Julius Marozas <marozas.julius@gmail.com>
 
 Permission is hereby granted, free of charge, to any
 person obtaining a copy of this software and associated
