+++ static/index.html
@@ -14,13 +14,11 @@
       <button id="reset">💣</button>
       <button id="blank">💀</button>
     </div>
-    <input type="range" id="tick-range" min="1" max="60">
-    <div id="fps"></div>
   </div>
 
   <canvas id="game-of-life-canvas"></canvas>
 
-  <script src="bootstrap.js"></script>
+  <script src="index.js"></script>
 </body>
 
 </html>
