+++ webpack.config.js
@@ -1,0 +1,39 @@
+const path = require('path');
+const CopyWebpackPlugin = require('copy-webpack-plugin');
+const WasmPackPlugin = require("@wasm-tool/wasm-pack-plugin");
+
+module.exports = {
+  entry: './www/bootstrap.ts',
+  output: {
+    path: path.resolve(__dirname, 'dist'),
+    filename: 'index.js',
+  },
+  devServer: {
+    contentBase: path.resolve(__dirname, 'dist'),
+  },
+  module: {
+    rules: [{
+        test: /\.tsx?$/,
+        use: 'ts-loader',
+        exclude: /node_modules/
+      },
+      {
+        test: /\.wasm$/,
+        type: "webassembly/experimental"
+      }
+    ]
+  },
+  resolve: {
+    extensions: ['.tsx', '.ts', '.js', '.wasm']
+  },
+  plugins: [
+    new CopyWebpackPlugin([
+      'static/index.html',
+      'static/styles.css',
+    ]),
+    new WasmPackPlugin({
+      crateDirectory: path.resolve(__dirname, '.'),
+      withTypeScript: true,
+    }),
+  ],
+};
