+++ www/bootstrap.ts
@@ -1,0 +1,1 @@
+import('./index').catch(e => console.error('Error importing `index.ts`:', e));
