--- www/package.json
@@ -1,25 +1,0 @@
-{
-  "name": "game-of-life",
-  "version": "0.1.0",
-  "main": "src/index.js",
-  "scripts": {
-    "build": "webpack --config webpack.config.js",
-    "start": "webpack-dev-server",
-    "deploy": "npm run build && cd .. && git subtree push --prefix www/dist origin gh-pages"
-  },
-  "keywords": [
-    "webassembly",
-    "wasm",
-    "rust",
-    "webpack"
-  ],
-  "author": "Julius Marozas <marozas.julius@gmail.com>",
-  "license": "MIT",
-  "devDependencies": {
-    "copy-webpack-plugin": "^4.6.0",
-    "webpack": "^4.28.3",
-    "webpack-cli": "^3.2.0",
-    "webpack-dev-server": "^3.1.14"
-  },
-  "dependencies": {}
-}
