--- www/public/index.html
@@ -1,26 +1,0 @@
-<!DOCTYPE html>
-<html>
-
-<head>
-  <meta charset="utf-8">
-  <title>Game of Life</title>
-  <link rel="stylesheet" href="styles.css">
-</head>
-
-<body>
-  <div class="sidebar">
-    <div class="controls">
-      <button id="play-pause" data-play="▶️" data-pause="⏸"></button>
-      <button id="reset">💣</button>
-      <button id="blank">💀</button>
-    </div>
-    <input type="range" id="tick-range" min="1" max="60">
-    <div id="fps"></div>
-  </div>
-
-  <canvas id="game-of-life-canvas"></canvas>
-
-  <script src="bootstrap.js"></script>
-</body>
-
-</html>
