--- www/public/styles.css
@@ -1,43 +1,0 @@
-body {
-  box-sizing: border-box;
-  position: absolute;
-  top: 0;
-  left: 0;
-  margin: 0;
-  padding: 8px;
-  width: 100%;
-  height: 100%;
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  justify-content: center;
-}
-
-.sidebar {
-  padding-right: 50px; 
-  display: flex;
-  flex-direction: column;
-}
-
-.controls {
-  text-align: center;
-  display: inline;
-}
-
-#fps {
-  white-space: pre;
-  font-family: monospace;
-}
-
-button {
-  background: none;
-  border: none;
-  margin: 0;
-  padding: 0;
-  cursor: pointer;
-  font-size: 18pt;
-}
-
-button:focus {
-  outline: 0;
-}
