--- www/src/utils.js
@@ -1,14 +1,0 @@
-CanvasRenderingContext2D.prototype.clear =
-  CanvasRenderingContext2D.prototype.clear ||
-  function(preserveTransform = false) {
-    if (preserveTransform) {
-      this.save();
-      this.setTransform(1, 0, 0, 1, 0, 0);
-    }
-
-    this.clearRect(0, 0, this.canvas.width, this.canvas.height);
-
-    if (preserveTransform) {
-      this.restore();
-    }
-  };
