--- www/webpack.config.js
@@ -1,14 +1,0 @@
-const CopyWebpackPlugin = require("copy-webpack-plugin");
-const path = require('path');
-
-module.exports = {
-  entry: "./src/bootstrap.js",
-  output: {
-    path: path.resolve(__dirname, "dist"),
-    filename: "bootstrap.js",
-  },
-  mode: "development",
-  plugins: [
-    new CopyWebpackPlugin(['public/index.html', 'public/styles.css'])
-  ],
-};
